+++ mods/GUIs/modsUI/instutils/iubpanel.cpp
@@ -915,8 +915,8 @@ void IUBPanel::update()
   emit statusMsg(tr("Updating %1 utility status...").arg(panelName),Qt::blue);
   sendCmdWait(modsIEHost[modsID],"UTIL STATUS",MODS_QUEUE_REQUEST);
   sendCmdWait(modsIEHost[modsID],"ESTATUS",MODS_QUEUE_REQUEST);
-  sendCmdWait(modsBCHost[modsID],"ESTATUS",MODS_QUEUE_REQUEST);
-  sendCmdWait(modsRCHost[modsID],"ESTATUS",MODS_QUEUE_REQUEST);
+  //sendCmdWait(modsBCHost[modsID],"ESTATUS",MODS_QUEUE_REQUEST);
+  //sendCmdWait(modsRCHost[modsID],"ESTATUS",MODS_QUEUE_REQUEST);
 
 }