+++ mods/GUIs/modsUI/mods.h
@@ -6,8 +6,8 @@
 //
 
 #define MODS_SESSION_NAME "MODS Control Panel" //!< Window Name for banner
-#define MODS_REV_NUMBER   "v3.1.1-archon"      //!< MODS UI Revision number
-#define MODS_REV_DATE     "2025 Dec 22"        //!< Revision date
+#define MODS_REV_NUMBER   "v3.2.0-archon"      //!< MODS UI Revision number
+#define MODS_REV_DATE     "2026 Jan 18"        //!< Revision date
 
 // Runtime files and paths
 
@@ -163,9 +163,9 @@
 
 #define MODS_MAX_MEASURES  32000 //!< Maximum number of measurments for housekeeping monitor data
 #define MODS_MAX_BOXTEMP    32.0 //!< Maximum box temperature before triggering cooling alarms [deg C]
-#define MODS_MAX_DEWPRES 1.0e-03 //!< Maximum dewar pressure before triggering alarms [torr]
-#define MODS_MAX_CCDTEMP   -90.0 //!< Maximum CCD temperature before triggering alarms [deg C]
-#define MODS_MAX_DEWTEMP  -165.0 //!< Maximum LN2 dewar tank temperature before triggering alarms [deg C]
+#define MODS_MAX_DEWPRES 1.0e-02 //!< Maximum dewar pressure before triggering alarms [torr]
+#define MODS_MAX_CCDTEMP   -50.0 //!< Maximum CCD temperature before triggering alarms [deg C]
+#define MODS_MAX_DEWTEMP   -50.0 //!< Maximum LN2 dewar tank temperature before triggering alarms [deg C]
 #define MODS_MIN_GLYPRES    -1.0 //!< Minimum pressure of the glycol cooling supply/return [psi-g], -1 disables
 #define MODS_ALARM_REPEAT      5 //!< Suppress repeat alarm messages every N times