+++ mods/GUIs/modsUI/releases.md
@@ -1,7 +1,7 @@
 # modsUI Release Notes
 Original Build: 2009 Feb 24
 
-Last Build: 2025 Dec 22
+Last Build: 2026 Jan 18
 
 ### modsUI version numbering
 
@@ -9,6 +9,12 @@ Last Build: 2025 Dec 22
  * Version 1.x - MODS1 monocular version for commissioning and monocular science with MODS1, using Qt4
  * Version 2.x - MODS1+2 binocular operations version
  * Version 3.x - MODS Archon CCD controller update version and AlmaLinux Port to Qt6
+
+## Version 3.2.0
+2026 Jan 18
+
+ * Removed all "ESTATUS" commands sent to RCHost and BCHost (not part of the modsCCD command set)
+ * Removed all "SAVECONFIG" commands (not part of modsCCD command set)
 
 ## Version 3.1.1
 2025 Dec 22